+++ days/09/day-09.hs
@@ -55,8 +55,8 @@ parseInput fileInput =
 updateLoc :: MoveDir -> Loc -> Loc
 updateLoc mvDir (r, c) =
   case mvDir of
-    'U' -> (r-1, c)
-    'D' -> (r+1, c)
+    'U' -> (r+1, c)
+    'D' -> (r-1, c)
     'L' -> (r, c-1)
     'R' -> (r, c+1)
     otherwise -> error "Invalid Move. Not in ['U','D','L','R']"
@@ -65,7 +65,7 @@ updateLoc mvDir (r, c) =
     * Move the head
     * Tail follows
         - is the FIRST step per move instruction,
-          the ONLY time Tail might not move?
+          the ONLY time Tail might not move? [NO! NOT TRUE!]
         - ex: H moves L then R and now H & T have the same Loc
             ie: T should not move
         - ex: H & T are at the same Loc, H moves
@@ -74,11 +74,27 @@ updateLoc mvDir (r, c) =
   =================================================================
 
     * Following above insight:
-        - Test for "must move" is only necessary on FIRST step
-        - Subsequent steps will mimic Head's movement.
-        - change doOneStep to doFirstStep
+        - [FALSE] Test for "must move" is only necessary on FIRST step [FALSE!]
+        - [FALSE] Subsequent steps will mimic Head's movement. [FALSE!]
+
+        [PROOF]
+        - ex: - H & T start at same Loc
+              - H moves R, T doesn't move
+              - H moves U, T doesn't move ( they are now diagonally apart )
+              - H moves D 3x, T doesn't move UNTIL 3rd step!
+
+              - cannot easily preprocess the data cancelling out L & R etc
+                as the moves might not immediately sequential...
+                as in the above example, the last line could be H moves L 3x
+
+        - NEW hypothesis:
+              - once T moves, then T will always move until MoveInstr completed
+
         - convert remaining steps to a List, then Set.fromList
         - calculate destination of Head and Tail
+
+    - changed doOneStep to doFirstStep
+    - changed doFirstStep to doStepWithCheck
 
   =================================================================
 
@@ -111,8 +127,9 @@ updateLoc mvDir (r, c) =
 mustMove :: Loc -> Loc -> Bool
 mustMove (r1,c1) (r2,c2) = abs (r1-r2) > 1 || abs (c1-c2) > 1
 
-doFirstStep :: MoveInstruction -> MoveRec -> MoveRec
-doFirstStep (mvDir, mvAmt) moveRec@(MoveRec {hLoc, tLoc, visited}) =
+-- check to see if T should follow H
+doStepWithCheck :: MoveDir -> MoveRec -> MoveRec
+doStepWithCheck mvDir moveRec@(MoveRec {hLoc, tLoc, visited}) =
   let newHLoc = updateLoc mvDir hLoc
   in
      if mustMove newHLoc tLoc
@@ -123,12 +140,19 @@ doFirstStep (mvDir, mvAmt) moveRec@(MoveRec {hLoc, tLoc, visited}) =
       else moveRec { hLoc = newHLoc }
 
 doFullMove :: MoveInstruction -> MoveRec -> MoveRec
-doFullMove moveInstr@(mvDir, mvAmt) moveRec@(MoveRec {hLoc, tLoc, visited}) =
-  -- mvAmt will always be > 1
-  let movesToMake = [1..(mvAmt-1)]
-      afterFirst  = doFirstStep moveInstr moveRec
+doFullMove (mvDir, mvAmt) moveRec@(MoveRec {hLoc, tLoc, visited}) =
+  -- mvAmt will always be >= 1
+  let afterFirst2 = undefined -- doStepWithCheck mvDir moveRec
+      newTLocs    = scanl (flip updateLoc) tLoc $ replicate (mvAmt-1) mvDir
+      newTLoc     = last newTLocs
+      -- newHLoc     = undefined
+      newHLoc     = foldl (flip updateLoc) hLoc $ replicate mvAmt mvDir
   in
-     error "scanl FTW?!"
+      error "afterFirst2 not used?! must start w/ updated hLoc & tLoc ..."
+      moveRec { hLoc = newHLoc
+              , tLoc = newTLoc
+              , visited = Set.union visited $ Set.fromList newTLocs
+              }
 
 main :: IO ()
 main = do
@@ -144,7 +168,7 @@ main = do
 
   let moveRec = MoveRec hLoc tLoc $ Set.fromList [tLoc]
 
-  putStrLn $ show $ doFirstStep (head moves) moveRec
+  putStrLn $ show $ doStepWithCheck (fst $ head moves) moveRec
 
   let moveRec1 = MoveRec (5,5) (5,5) $ Set.fromList [(5,5)]
   let moveRec2 = MoveRec (5,5) (4,5) $ Set.fromList [(4,5)]
@@ -154,13 +178,15 @@ main = do
   putStrLn $ show $ head moves
   putStrLn ""
   putStrLn $ show $ moveRec1
-  putStrLn $ " ==> " ++ show (doFirstStep (head moves) moveRec1)
+  putStrLn $ " ==> " ++ show (doStepWithCheck (fst $ head moves) moveRec1)
   putStrLn ""
   putStrLn $ show $ moveRec2
-  putStrLn $ " ==> " ++ show (doFirstStep (head moves) moveRec2)
+  putStrLn $ " ==> " ++ show (doStepWithCheck (fst $ head moves) moveRec2)
   putStrLn ""
   putStrLn $ show $ moveRec3
-  putStrLn $ " ==> " ++ show (doFirstStep (head moves) moveRec3)
+  putStrLn $ " ==> " ++ show (doStepWithCheck (fst $ head moves) moveRec3)
 
-  putStrLn $ show (doFullMove ('D',10) (MoveRec (0,0) (1,1) Set.empty))
-  
+  let moveRec4 = MoveRec (0,0) (1,1) $ Set.fromList [(1,1)]
+  putStrLn ""
+  putStrLn $ show $ moveRec4
+  putStrLn $ " ==> " ++ show (doFullMove ('U', 10) moveRec4)